Moved pages, assets and components to src folder
