+++ Dockerfile
@@ -9,4 +9,4 @@ FROM nginx:1.17.8-alpine
 LABEL maintainer="Alejandro Celaya <alejandro@alejandrocelaya.com>"
 RUN rm -r /usr/share/nginx/html && rm /etc/nginx/conf.d/default.conf
 COPY --from=node /shlink-website/out /usr/share/nginx/html
-COPY assets/nginx.conf /etc/nginx/conf.d/default.conf
+COPY src/assets/nginx.conf /etc/nginx/conf.d/default.conf
