+++ src/components/FakeBrowser.tsx
