+++ src/components/SectionHeader.tsx
