+++ src/components/Terminal.tsx
