+++ test/components/Feature.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, cleanup } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import Feature from '../../components/Feature';
+import Feature from '../../src/components/Feature';
 
 describe('<Feature />', () => {
   const theTitle = 'The title';
