+++ test/components/SectionHeader.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, cleanup } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import SectionHeader from '../../components/SectionHeader';
+import SectionHeader from '../../src/components/SectionHeader';
 
 describe('<SectionHeader />', () => {
   afterEach(cleanup);
